refactor: get rid of mattermost auth layer
