+++ server/services/store/sqlstore/compliance.go
@@ -34,7 +34,7 @@ func (s *SQLStore) getBoardsForCompliance(db sq.BaseRunner, opts model.QueryBoar
 	}
 	defer s.CloseRows(rows)
 
-	boards, err := s.boardsFromRows(rows)
+	boards, err := s.boardsFromRows(rows, false)
 	if err != nil {
 		return nil, false, err
 	}
