+++ server/services/store/sqlstore/messages.go
@@ -1,0 +1,72 @@
+package sqlstore
+
+import (
+	"github.com/mattermost/mattermost-plugin-boards/server/model"
+	mmModel "github.com/mattermost/mattermost/server/public/model"
+	"github.com/mattermost/mattermost/server/public/shared/mlog"
+)
+
+func (s *SQLStore) getBoardsBotID() (string, error) {
+	var boardsBotID string
+	if boardsBotID == "" {
+		var err error
+		boardsBotID, err = s.servicesAPI.EnsureBot(model.FocalboardBot)
+		if err != nil {
+			s.logger.Error("failed to ensure boards bot", mlog.Err(err))
+			return "", err
+		}
+	}
+	return boardsBotID, nil
+}
+
+func (s *SQLStore) SendMessage(message, postType string, receipts []string) error {
+	botID, err := s.getBoardsBotID()
+	if err != nil {
+		return err
+	}
+
+	for _, receipt := range receipts {
+		channel, err := s.servicesAPI.GetDirectChannel(botID, receipt)
+		if err != nil {
+			s.logger.Error(
+				"failed to get DM channel between system bot and user for receipt",
+				mlog.String("receipt", receipt),
+				mlog.String("user_id", receipt),
+				mlog.Err(err),
+			)
+			continue
+		}
+
+		if err := s.PostMessage(message, postType, channel.Id); err != nil {
+			s.logger.Error(
+				"failed to send message to receipt from SendMessage",
+				mlog.String("receipt", receipt),
+				mlog.Err(err),
+			)
+			continue
+		}
+	}
+	return nil
+}
+
+func (s *SQLStore) PostMessage(message, postType, channelID string) error {
+	botID, err := s.getBoardsBotID()
+	if err != nil {
+		return err
+	}
+
+	post := &mmModel.Post{
+		Message:   message,
+		UserId:    botID,
+		ChannelId: channelID,
+		Type:      postType,
+	}
+
+	if _, err := s.servicesAPI.CreatePost(post); err != nil {
+		s.logger.Error(
+			"failed to send message to receipt from PostMessage",
+			mlog.Err(err),
+		)
+	}
+	return nil
+}
